+++ packages/bpk-tokens/gulpfile.babel.js
@@ -27,6 +27,7 @@ import theo from 'theo';
 import bpkScss from './formatters/bpk.scss';
 import bpkDefaultScss from './formatters/bpk.default.scss';
 import bpkEs6Js from './formatters/bpk.es6.js';
+import bpkReactNativeEs6Js from './formatters/bpk.react.native.es6.js';
 import bpkCommonJs from './formatters/bpk.common.js';
 import bpkAndroid from './formatters/bpk.android.xml';
 
@@ -38,17 +39,19 @@ const OUTPUT_MAP = {
     'common.js',
     'es6.js',
   ],
-  ios: ['ios.json'],
-  android: ['android.xml'],
+  ios: ['ios.json', 'react.native.es6.js', 'raw.json'],
+  android: ['android.xml', 'react.native.es6.js', 'raw.json'],
 };
 
 const WEB_OUTPUTS = _(OUTPUT_MAP.web).map(format => ({ transform: 'web', format }));
+const IOS_OUTPUTS = _(OUTPUT_MAP.ios).map(format => ({ transform: 'ios', format }));
 
 // Theo config
 
 theo.registerFormat('scss', bpkScss);
 theo.registerFormat('default.scss', bpkDefaultScss);
 theo.registerFormat('es6.js', bpkEs6Js);
+theo.registerFormat('react.native.es6.js', bpkReactNativeEs6Js);
 theo.registerFormat('common.js', bpkCommonJs);
 theo.registerFormat('android.xml', bpkAndroid);
 
@@ -86,8 +89,12 @@ const makeConverter = platform => (options, done) => {
 
 gulp.task('tokens', ['clean', 'lint'], (done) => {
   const webConverter = makeConverter('web');
+  const iosConverter = makeConverter('ios');
 
-  async.each(WEB_OUTPUTS, webConverter, done);
+  async.parallel([
+    callback => async.each(WEB_OUTPUTS, webConverter, callback),
+    callback => async.each(IOS_OUTPUTS, iosConverter, callback),
+  ], done);
 });
 
 gulp.task('default', ['tokens']);