+++ packages/bpk-tokens/tokens/ios/base.react.native.es6.js
@@ -1,0 +1,239 @@
+/*
+ * 
+ * Backpack - Skyscanner's Design System
+ * 
+ * Copyright 2017 Skyscanner Ltd
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * 
+ */
+export const colorWhite = "rgb(255, 255, 255)";
+export const colorBlue50 = "rgb(225, 244, 248)";
+export const colorBlue100 = "rgb(203, 238, 245)";
+export const colorBlue200 = "rgb(176, 228, 238)";
+export const colorBlue300 = "rgb(127, 215, 232)";
+export const colorBlue400 = "rgb(64, 196, 223)";
+export const colorBlue500 = "rgb(0, 178, 214)";
+export const colorBlue600 = "rgb(0, 157, 189)";
+export const colorBlue700 = "rgb(0, 140, 168)";
+export const colorBlue800 = "rgb(0, 117, 140)";
+export const colorBlue900 = "rgb(0, 85, 103)";
+export const colorGray50 = "rgb(243, 242, 245)";
+export const colorGray100 = "rgb(230, 228, 235)";
+export const colorGray200 = "rgb(204, 201, 212)";
+export const colorGray300 = "rgb(178, 174, 189)";
+export const colorGray400 = "rgb(154, 149, 167)";
+export const colorGray500 = "rgb(129, 123, 143)";
+export const colorGray600 = "rgb(105, 97, 121)";
+export const colorGray700 = "rgb(82, 76, 97)";
+export const colorGray800 = "rgb(59, 52, 75)";
+export const colorGray900 = "rgb(37, 32, 51)";
+export const colorGreen50 = "rgb(223, 247, 236)";
+export const colorGreen100 = "rgb(203, 245, 226)";
+export const colorGreen200 = "rgb(175, 237, 209)";
+export const colorGreen300 = "rgb(128, 232, 185)";
+export const colorGreen400 = "rgb(64, 222, 151)";
+export const colorGreen500 = "rgb(0, 215, 117)";
+export const colorGreen600 = "rgb(0, 189, 104)";
+export const colorGreen700 = "rgb(0, 168, 93)";
+export const colorGreen800 = "rgb(0, 140, 77)";
+export const colorGreen900 = "rgb(0, 102, 56)";
+export const colorRed50 = "rgb(252, 242, 242)";
+export const colorRed100 = "rgb(255, 214, 213)";
+export const colorRed200 = "rgb(255, 187, 186)";
+export const colorRed300 = "rgb(255, 150, 148)";
+export const colorRed400 = "rgb(254, 116, 113)";
+export const colorRed500 = "rgb(255, 84, 82)";
+export const colorRed600 = "rgb(235, 66, 63)";
+export const colorRed700 = "rgb(222, 50, 47)";
+export const colorRed800 = "rgb(204, 31, 29)";
+export const colorRed900 = "rgb(168, 3, 0)";
+export const colorYellow50 = "rgb(255, 249, 230)";
+export const colorYellow100 = "rgb(255, 243, 207)";
+export const colorYellow200 = "rgb(255, 236, 184)";
+export const colorYellow300 = "rgb(255, 225, 140)";
+export const colorYellow400 = "rgb(255, 207, 74)";
+export const colorYellow500 = "rgb(255, 187, 0)";
+export const colorYellow600 = "rgb(240, 176, 0)";
+export const colorYellow700 = "rgb(225, 165, 0)";
+export const colorYellow800 = "rgb(194, 142, 0)";
+export const colorYellow900 = "rgb(156, 114, 0)";
+export const spacingSm = 4;
+export const spacingMd = 8;
+export const spacingBase = 16;
+export const spacingLg = 24;
+export const spacingXl = 32;
+export const spacingXxl = 40;
+export const fontSizeXs = 11;
+export const fontSizeSm = 13;
+export const fontSizeBase = 15;
+export const fontSizeLg = 17;
+export const fontSizeXl = 20;
+export const fontSizeXxl = 34;
+export const lineHeightXs = 13;
+export const lineHeightSm = 18;
+export const lineHeightBase = 20;
+export const lineHeightLg = 22;
+export const lineHeightXl = 24;
+export const lineHeightXxl = 41;
+export const letterSpacingXs = 0.07;
+export const letterSpacingSm = -0.08;
+export const letterSpacingBase = -0.24;
+export const letterSpacingLg = 0;
+export const letterSpacingXl = 0.38;
+export const letterSpacingXxl = 0.41;
+export const fontWeightNormal = "400";
+export const fontWeightSemibold = "600";
+export const fontWeightBold = "700";
+export const fontWeightHeavy = "800";
+export const textXsFontSize = 11;
+export const textXsFontWeight = "400";
+export const textXsLineHeight = 13;
+export const textXsLetterSpacing = 0.07;
+export const textSmFontSize = 13;
+export const textSmFontWeight = "400";
+export const textSmLineHeight = 18;
+export const textSmLetterSpacing = -0.08;
+export const textBaseFontSize = 15;
+export const textBaseFontWeight = "400";
+export const textBaseLineHeight = 20;
+export const textBaseLetterSpacing = -0.24;
+export const textLgFontSize = 17;
+export const textLgFontWeight = "400";
+export const textLgLineHeight = 22;
+export const textLgLetterSpacing = 0;
+export const textXlFontSize = 20;
+export const textXlFontWeight = "400";
+export const textXlLineHeight = 24;
+export const textXlLetterSpacing = 0.38;
+export const textXxlFontSize = 34;
+export const textXxlFontWeight = "700";
+export const textXxlLineHeight = 41;
+export const textXxlLetterSpacing = 0.41;
+export const colors = {
+colorWhite,
+colorBlue50,
+colorBlue100,
+colorBlue200,
+colorBlue300,
+colorBlue400,
+colorBlue500,
+colorBlue600,
+colorBlue700,
+colorBlue800,
+colorBlue900,
+colorGray50,
+colorGray100,
+colorGray200,
+colorGray300,
+colorGray400,
+colorGray500,
+colorGray600,
+colorGray700,
+colorGray800,
+colorGray900,
+colorGreen50,
+colorGreen100,
+colorGreen200,
+colorGreen300,
+colorGreen400,
+colorGreen500,
+colorGreen600,
+colorGreen700,
+colorGreen800,
+colorGreen900,
+colorRed50,
+colorRed100,
+colorRed200,
+colorRed300,
+colorRed400,
+colorRed500,
+colorRed600,
+colorRed700,
+colorRed800,
+colorRed900,
+colorYellow50,
+colorYellow100,
+colorYellow200,
+colorYellow300,
+colorYellow400,
+colorYellow500,
+colorYellow600,
+colorYellow700,
+colorYellow800,
+colorYellow900,
+};
+export const spacings = {
+spacingSm,
+spacingMd,
+spacingBase,
+spacingLg,
+spacingXl,
+spacingXxl,
+};
+export const typesettings = {
+fontSizeXs,
+fontSizeSm,
+fontSizeBase,
+fontSizeLg,
+fontSizeXl,
+fontSizeXxl,
+lineHeightXs,
+lineHeightSm,
+lineHeightBase,
+lineHeightLg,
+lineHeightXl,
+lineHeightXxl,
+letterSpacingXs,
+letterSpacingSm,
+letterSpacingBase,
+letterSpacingLg,
+letterSpacingXl,
+letterSpacingXxl,
+};
+export const fontWeights = {
+fontWeightNormal,
+fontWeightSemibold,
+fontWeightBold,
+fontWeightHeavy,
+textXsFontWeight,
+textSmFontWeight,
+textBaseFontWeight,
+textLgFontWeight,
+textXlFontWeight,
+textXxlFontWeight,
+};
+export const fontSizes = {
+textXsFontSize,
+textSmFontSize,
+textBaseFontSize,
+textLgFontSize,
+textXlFontSize,
+textXxlFontSize,
+};
+export const lineHeights = {
+textXsLineHeight,
+textSmLineHeight,
+textBaseLineHeight,
+textLgLineHeight,
+textXlLineHeight,
+textXxlLineHeight,
+};
+export const letterSpacings = {
+textXsLetterSpacing,
+textSmLetterSpacing,
+textBaseLetterSpacing,
+textLgLetterSpacing,
+textXlLetterSpacing,
+textXxlLetterSpacing,
+};